Split walker/expressions_p2.rs at the 800-line cap into expressions_p3.rs
